requirements.txt
# Web framework
fastapi==0.135.1
uvicorn[standard]==0.41.0

# MongoDB
pymongo==4.16.0
beanie==2.0.1

# LangChain + LangGraph
langchain==1.2.10
langchain-core==1.2.17
langgraph==1.0.10

# Validation
pydantic==2.12.5
pydantic-settings==2.13.1

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
testcontainers[mongodb]==4.14.1
python-dotenv==1.2.2
markdown==3.10.2